project/intro_project/requirements.txt
Django==1.11.11
Pillow==8.4.0
sqlparse==0.2.4
numpy==1.19.5
scikit-learn==0.24.1
psycopg2-binary==2.8.6
tensorflow==2.6.2